simply run generate-samples.sh

+++ samples/client/petstore/csharp-netcore/OpenAPIClient-generichost-net6.0-nrt/src/Org.OpenAPITools/Model/Banana.cs
@@ -117,7 +117,7 @@ namespace Org.OpenAPITools.Model
                     {
                         case "lengthCm":
                             if (utf8JsonReader.TokenType != JsonTokenType.Null)
-                                lengthCm = utf8JsonReader.GetDecimal();
+                                lengthCm = utf8JsonReader.GetInt32();
                             break;
                         default:
                             break;
